requirements.txt
django
numpy
pandas
djangorestframework
django-cors-headers
joblib
scikit-learn
DecisionTreeClassifier